Allow an override of the library from the manifest that imports a git/svn module.
Override the library variable by specifying the git/svn module as a tuple (url, library) instead of just the url string.
The module path (checkout directory) is modified by appending the repository basename with the library. 
Backwards compatibility with plain url is maintained.
This allows multiple versions of the same repository to be imported into the project, with them separated into different library namespaces.